Update 10_edb_resource_manager.mdx
We might need to clarify some part of the docs on `cpu_rate_limit`. Docs said:
- "the value assigned to cpu_rate_limit should typically be less than or equal to 1."
- and also "The valid range of the cpu_rate_limit parameter is 0 to 1.67772e+07"

The value could be more than 1.0 on a multicore system. So for example 1.0 would mean to used 100% of one CPU core, and 2.0 would mean to use 100% of 2 CPUs. That understanding was confirmed on the `eng-db-server` slack channel (https://edb.slack.com/archives/CDTA6PCJV/p1628781278238200) 

Please feel free to change the wording if needed.