doc(UPM-2321): Document CNP metrics exposed by UPM
List the CNP metrics exposed by UPM.

Also provide some guidance on using those metrics and on the
structure of metrics and logs entries.

Note that this documentation change contains a section that is generated
by a script. The script is indended to be hosted in the upm-substrate
repo. It doesn't seem practical to add the script here and have it
re-generate the automatically generated section on every run, so
updating it is expected to be part of the UPM release process for now.