more information for efm slot update property
The database itself will enforce a unique slot name for all of the standbys, but the name specified in the primary's configuration isn't used by the database until another node is promoted, so the database doesn't care. If the primary uses the same name that a standby is using, things won't work as expected.