Adjust dependencies of pgBackrest in RHEL 7
In a support ticket (83285), a customer followed these directions and installed `postgresql-libs` manually before pgbackrest. Later on, that package conflicted with other dependencies of other packages, in the typical collision on `/usr/lib64/libpq.so`.

In RHEL 7, `postgresql-libs` brings libpq 9.2 of Red Hat, which is not a supported version and is also not tested against any of our packages. In RHEL 8, that package does not exist; instead, RHEL 8 provides `libpq` in the appstream repository.

In both cases, the package provided by PGDG is called `libpq5`, which is the right expectation to set when installing `pgbackrest` also from PGDG.

Additionally, if other packages have been installed before and already pulled these dependencies, then maybe these libraries won't be pulled in when installing `pgbackrest`, so this also adjusts the wording slightly to reflect that.