change to efm standby priority description
The original text here says that the user may reset the value, which isn't something we should direct the user to do (or even explain how they *could* do it). The text is supposed to say that it's something that can happen. The original text in the workspace didn't say that the user could do it, so this was messed up somewhere later.