Restructure of the Patroni documentation
We've moved to a schema similar to other tools like EFM.
The installation is separated in two sections, one for Debian based
distributions (which includes Ubuntu) and another for RHEL (which
works for Alma Linux, Rocky and CentOS 7). We provided instructions
for installing PostgreSQL and EPAS for each of the two families of
distributions.
Installation of etcd was left in a separate section.
All the configuration of Patroni, watchdog, etc. was moved to a
separate section.
Connection configuration and HAProxy is now in a "Connecting"
section.
Also changed some code blocks, leaving with `bash` only the ones
that are commands that can be copied and pasted in a shell.

This will very likely require another round of reviews and
probably also changes.

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>