Clean up removed files and move diagram to the supported platforms section
We left over the Installing_for_epas.mdx and installing_for_postgresql.mdx
files from previous commits (we rearranged the sections in the
documentation).

Moved the diagram with the architecture to the supported platforms section
and a paragraph describing the architecture

Also added a link to the platform compatibility web page.

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>