Add a bullet point for HAProxy in the supported arch in the index
Added a line with HAProxy in the index page from Patroni docs,
expaning a bit on what we support, what we strongly suggest.

Later we might want to expand on why it's importnat to have HAProxy
and why two nodes with HAProxy. Or show other configurations,
like having HAProxy on the application node, or on the database nodes.

Reported by: Bobby Bissett

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>