efm_timeout_descriptions
The description for remote.timeout was misleading in that someone could infer that agents generally monitor remote databases. Agents only do this (communicate in general, that is) when there's some reason to, e.g. during a potential failover. They also send messages like "shutdown for switchover" to the primary, "reconfigure to follow the new primary" to standbys, "prepare to shutdown," etc. We could put in more examples if anyone thinks that would be more helpful in this section.

I'm suggesting we change the node.timeout one because an agent waiting "for a response" implies the agent sent a request to another. While "response" isn't technically NOT true, the real thing an agent is waiting on is a heartbeat message or any kind of communication from another agent before deciding "ok, it's gone."