Adjust indentation of a !!! Important section
Currently, that section shows up literally as `!!! Important Alternatively, you can list namespaces using a label selector, as explained in` in the documentation.
I believe that might be caused by the block being 1 space further in, when compared to the rest of the 3. section, so this change tries to handle that.