Noted PG/EPAS 10 and Oracle 10g no longer supported
Added a row in the table to note that PostgreSQL 10 and EPAS 10 are no longer supported for use with Replication Server.  Also, added a row to note that Oracle 10g is no longer officially supported for use with Replication Server.  Also made a tweak to the wording for PostgreSQL 15 support.