Update azure_ad.mdx
Make below steps clear:
Select Test Connection. If the connection is successful, select Sign in to BigAnimal to complete the setup process in the BigAnimal portal.
For any further BigAnimal sign in, select Sign in to your BigAnimal account instead of the button of Sign in for Azure Marketplace users.