Synchronous Replication link name
The link pointer name is "Synchronous Replication" however when we click on the link the page is generic and there is only one time when the replication lag is referred. So essentially there is not much information about the replication lag and if we are referring to the one which exists on that page then that is under Standby Replicas, so it better to name the link to the page as Standby Replicas and not as Synchronous Replication.

This is not a pressing need just for the sack of consistency, as a user when I tried it myself, I thought I have a broken link as I was expecting information on Replication Lag and the link was pointing me to Synchronous Replication I thought there will be dedicated section for this one.