Update installing_the_sql_profiler_plugin.mdx
Removed the statement that you don't need to install with EPAS as it seems to be untrue (and there are instructions for doing exactly that below). I think this page will need much bigger changes very soon as the server team are releasing a new version that is installed differently.