Update activating_regions.mdx
As of a few weeks ago, activation of a region prior to cluster creation (or restoration into a region) is now a requirement.  Customers can no longer activate a region at the time of cluster creation, so made a few minor changes.  Please reach out if you have any questions.  Thanks!