Add information about how efm interprets ping
Adding a note that the important part about pinging the VIP is the exit code of the ping command. This comes from a customer replacing /bin/ping with something else that always exited with zero even though the human-readable output. In that case EFM always thought the address being pinged was reachable, breaking failover when using a VIP.