Added troubleshooting section headings and other updates
Added section headings to better organize the troubleshooting text.  Also, rearranged the presentation the hostname issue.  Finally, removed the note about the extractor not extracting all queries and changed to a new troubleshooting heading and updated the text to reflect that the issue is with the content that Oracle makes available in the V$SQL view.