EPAS 13-15: Fix some instructions on how to edit service files
This not only removes the usage of the deprecated `.include` feature, but also corrects the instructions that wrongfully directed customers towards editing files that were not meant to be edited.

It only covers some of the instructions that edit service files. More
fixing is necessary afterwards in other sections of the documentation.