Update 05_using_efm.mdx
This changes the instructions on how to customize systemd service files. The instructions were directing the reader to edit files in /usr/lib/systemd/system, which is not the adequate location. Whenever a system administrator edits service files, they should always do so in /etc/systemd/system, either with systemctl edit or editing the files directly, but never editing anything in /lib or /usr.