pg_extensions/ASP/using: improve examples
The example for refdata was improved to show the different locking
pattern that replaces the row locks for a table lock, leading to a
stronger lock when the reference table is changed, but a lower overhead
when the referenced rows are accessed. The demonstration uses
pgrowlocks, which is available for heap, and will be available in a
future version of our custom TAMs.

The example for autocluster was maintained.

The advanced example that uses both TAMs was improved to show the
effect that autocluster has on the query plan, leading to fewer pages
touched. Previously it showed the ctid of the rows, but that is already
displayed in the autocluster example, so showing the different query
plans fills a gap in the overall explanation of autocluster.

The examples were reworked to not need external data, through the usage
of artificial data generated on the fly.