removed text that was not needed
The db.bin property is just one of many needed properties, so calling it out here in a note doesn't make sense. My guess is that we added that property some time back and an upgrade required setting it, but by now any installation that a user is upgrading from will already have it set.

The unit file comment I removed doesn't make sense here, and I don't know where it might have come from. When upgrading, the service name doesn't change, and there would be a new unit file anyway. This might be related to the multiple efm agents on a single node section, but doesn't make sense here that I can see.