PGD: rearrange internal functions reference
This page starts off with the functions in alphabetical order, which
breaks down part-way through, then there's a subsection titled
"Task Manager Functions", which contains a couple of taskmanager
functions, then a bunch of unrelated functions which were presumably
just added to the end of the page.

As the functions start off in alphabetical order, a human scanning
the list will tend to assume a function is not listed if it's
not found where expected alphabetically.

This commit sorts the non-taskmanager functions alphabetically to
make the page more usable.

It would probably make sense to group functions by purpose (view
helpers, KV store, consensus, backwards compatibility) etc. but
that is a larger refactoring.