Added more information about upgrade-conf source flag
This removes a little information from the general 'efm' utility page and adds a note to see the upgrade page for more info instead. Hopefully this is better than having incomplete information spread out across pages. On the upgrade page there is now a whole section on the -source parameter and when root privileges are not required to run it.

There is also a note about what happens when files already exist in the target location.

The above changes are in response to a customer confused about when root privileges were required or not, and what happened when files are upgraded into the same directory as the source files (source files renamed). Any suggestions are welcome -- the "summary" could possibly be a table but that might be overkill.