Add more information about efm startup log
This changes adds notes in two places to tell the user to check the startup log if there is a problem starting efm. Too often the user only checks 'systemctl status' output, which does not contain the problem.

(I'd like to get rid of the startup log because users forget it's there and it's not needed any more. I don't know when it will happen though: https://enterprisedb.atlassian.net/browse/EFM-1619)