Update 05_database_version_policy.mdx
Changed the explicit reference to "Postgres version 11" to "Postgres version X" since the line item applies to the two versions of PG shown in the columns - versions 11 and 12. If I misunderstood the intent, then please ignore this PR, but as I read it, the reference to only PG 11 was not accurate in the table.