fix autocluster example
The example had timestamptz as datatype and inserted only hour:minute
unquoted.

Besides the obvious quoting, we should either change the data to match
the datatype, or the datatype to match the hour:minute data. Using
full timestamptz probably wouldn't improve the example, which is well
served by simple data. So using `time` as column datatype is adequate.