Removed Prerequiste section and other minor changes
Removed the Prerequiste section. This info was in the quickstart guide created for EDB users, which was a standalone document.  However, as part of the online EDB Docs documentation, this info is not needed, since it is already covered elsewhere in the documentation.

Made some slight changes to the enabling and disabling instructions.