PGD: clarify table access method extension usage
This commit adds some cross-references to the "EDB Advanced Storage
Pack" documentation, which contains further information about the
respective table access methods.

It also clarifies the procedure for installing the extensions on a PGD
cluster, in particular emphasising that "shared_preload_libraries"
needs to be configured on each node, whereas the "CREATE EXTENSION"
command only needs to be executed on one node.

BDR-5140.