PGD: improve description of "transaction_id" pseudo-GUC
This is only queryable as a reported GUC via libpq's PQparameterStatus()
function, and is never accessible at SQL level. Rewrite the description
to clarify this, linking to a helpful code example while we're at it.

Also rework the introductory section to make it clearer that one of the
GUC-like objects about to be described is not query-able via SQL.

DOCS-1028.