pgd: clarify joining nodes with non-default extensions
If a node is joined with a user who does not have superuser permissions,
the presence of any extensions on the source node which require
superuser permissions to install, will cause the join operation to
fail.

Clarify that such extensions will need to be created manually on the
joining node if the user is not a superuser.

BDR-5519.