Added a paragraph about the Promote Status section of cluster status output
This leaves the existing info mostly alone, but adds an explanation of the Promote Status output and why it's there. Users sometimes think this is meant to check if replication is working, or want more information listed. The info that efm collects here is all related to whether or not a promotion would be safe (or a switchover allowed) and how a standby would be picked. It's not meant to be the place a user monitors replication. The existing text goes into detail a little about it, but this new text explains why this particular info is here.