Removed Windows Server 2019 from link
The link to the Windows installation page, listed Windows Server 2019, 2022, and Windows 11.  Removed Windows Server 2019 from the list and adjusted the rest of the link label accordingly since Windows Server 2019 went end of life over a year ago and we no longer support EDB*Plus or any of our products on it.