[EFM] 5.0 changes (Friday release) (#6594)
* Changes for EFM 5.0

See story for status on which changes are in this PR already and which are still waiting:
https://enterprisedb.atlassian.net/browse/EFM-1730

This first commit makes general version bumps across the docs without any new content yet. A placeholder release notes file has been added; am pushing now to get some help with figuring out the release notes ordering and various "version 4" bits in the generated pages that I don't know how to find.

* Changes for EFM 5.0

See story for status on which changes are in this PR already and which are still waiting:
https://enterprisedb.atlassian.net/browse/EFM-1730

This first commit makes general version bumps across the docs without any new content yet. A placeholder release notes file has been added; am pushing now to get some help with figuring out the release notes ordering and various "version 4" bits in the generated pages that I don't know how to find.

* Fix build (no unescaped less-than symbols)

* ...missed one

* Updated 5.0 link in release notes index. Was still pointing at 4.10. Also removed the "5" from "...describes the latest version of Failover Manager" since the 4 or 5 isn't really important and I want to try to make it clear the docs cover the latest version no matter what it is, and changes are in the release notes.

* Fix a 4.x startup log reference that shouldn't have been changed to 5.x.

* Fixed all startup log references (hopefully).

* Fixed "xlog" text in json output.

* Changing the efm version directory from 4 to 5. No other changes in this commit.

* Fixed a broken link.

Helpful error messages and "npm run links:check" rule!

* We ran out of names following the existing convention, so I've bumped all the numbers way up.

This did involve fixing a link (checked with "npm run links:check") in a pdg page that referenced an efm release.

* Updated/removed some more startup log references and fixed the general logging page.

Added a big warning about not using TRACE level.

* Updated auto.resume.period information.

* Update text for detach.on.agent.failure property.

* Add new property backup.wal.

* Add new property check.num.sync.period.

This adds the property information, along with changing the wal_sender_timeout note to say it doesn't apply to efm versions 5 and up.

* Document "DOWN" cluster members.

This adds a note to the cluster-status and 'efm reset-members' pages about DOWN nodes.

* Adding the new efm create-standby command.

* Update to release notes.

Added the 5.0 release notes, fixed a stray apostrophe in the 4.8 notes, and updated the 5.0 release date.

* Remaining 4.x references and minor edits

* Added Addresses column to all release notes

* minor edits to 5.o rn

* Removed Ubuntu 20, updated config.yaml and added _redirects static path to catch /4/ traffic

* minor rn edits

* Added some text about detach.on.agent.failure property.

* rephrase detach_on_agent_failure

---------

Co-authored-by: Josh Heyer <josh.heyer@enterprisedb.com>
Co-authored-by: gvasquezvargas <gaby.vasquez@enterprisedb.com>