Have gone through 5.1 release on Jira and added all items
Info is here:
https://enterprisedb.atlassian.net/projects/EFM/versions/15314/tab/release-report-all-issues

Everything is in the docs now, though many of them are still marked
as "TODO" to write the rest. Everything is included in this draft
of the release notes (still need to write a description for one of
them, but wanted to get the jira release pass done).