add info about 5.0 version of create-standby
In 5.0 when the 'efm create-standby' command was introduced, it required superuser privileges to run since it needed to be able to start the db as a service without using the normal efm scripts. In 5.1 this was fixed so that it can be run as the 'efm' user, and the documentation was changed to remove the requirement text, but people will still get caught up on this if using 5.0.