Remove the db maintenance task section
This removes the section about performing db maintenance tasks from the
EFM upgrade page. This information wasn't relevant to upgrading efm,
and was causing problems with Rai parsing. The only part of this relevant
to efm was that someone should stop the efm agent before stopping the db
for maintenance -- the rest (e.g. order of nodes to work on) is specific
to postgres and is documented elsewhere.