PostGIS: update supported platforms
This adds v18 to the supported platforms, for PostGIS 3.6, which added
support for v18.

It also removes v13 from the table, since it's out of our main supported
platforms page and went EOL last month.

Additionally, the oldest version of PostGIS available both on our
repositories and in the release notes of our documentation is 3.1. So
this also removes the older versions from the supported matrix. Notice
that the upgrade steps still cover postgis versions 2.x, which is
reasonable since we want to continue guiding customers through the
upgrade process into supported versions, just not claim 2.x on its own
is supported anymore.